Moved shared functionality to FarmOpenExistingVMBase
